Move getList to class ProjectsList
